Flakify _ctypes.py

+++ datoviz/_ctypes.py
@@ -30,7 +30,7 @@ try:
     from numpy.ctypeslib import ndpointer as ndpointer_
 except ImportError:
     float32 = object
-    raise ImportError("NumPy is not available")
+    raise ImportError('NumPy is not available')
 
 
 # ===============================================================================
@@ -45,18 +45,18 @@ faulthandler.enable()
 # ===============================================================================
 
 PLATFORMS = {
-    "Linux": "linux",
-    "Darwin": "macos",
-    "Windows": "windows",
+    'Linux': 'linux',
+    'Darwin': 'macos',
+    'Windows': 'windows',
 }
 PLATFORM = PLATFORMS.get(platform.system(), None)
 
 LIB_NAMES = {
-    "linux": "libdatoviz.so",
-    "macos": "libdatoviz.dylib",
-    "windows": "libdatoviz.dll",
+    'linux': 'libdatoviz.so',
+    'macos': 'libdatoviz.dylib',
+    'windows': 'libdatoviz.dll',
 }
-LIB_NAME = LIB_NAMES.get(PLATFORM, "")
+LIB_NAME = LIB_NAMES.get(PLATFORM, '')
 
 FILE_DIR = pathlib.Path(__file__).parent.resolve()
 
@@ -67,12 +67,12 @@ LIB_PATH = DATOVIZ_DIR / LIB_NAME
 
 # Development paths: the libraries are in build/ and libs/
 if not LIB_PATH.exists():
-    DATOVIZ_DIR = (FILE_DIR / "../build/").resolve()
-    LIB_DIR = (FILE_DIR / f"../libs/vulkan/{PLATFORM}/").resolve()
+    DATOVIZ_DIR = (FILE_DIR / '../build/').resolve()
+    LIB_DIR = (FILE_DIR / f'../libs/vulkan/{PLATFORM}/').resolve()
     LIB_PATH = DATOVIZ_DIR / LIB_NAME
 
 if not LIB_PATH.exists():
-    raise RuntimeError(f"Unable to find `{LIB_PATH}`.")
+    raise RuntimeError(f'Unable to find `{LIB_PATH}`.')
 
 
 # ===============================================================================
@@ -83,17 +83,18 @@ assert LIB_PATH.exists()
 try:
     dvz = ctypes.cdll.LoadLibrary(str(LIB_PATH))
 except Exception as e:
-    print(f"Error loading library at {LIB_PATH}: {e}")
+    print(f'Error loading library at {LIB_PATH}: {e}')
     exit(1)
 
 # on macOS, we need to set the VK_DRIVER_FILES environment variable to the path to the MoltenVK ICD
-if PLATFORM == "macos":
-    os.environ["VK_DRIVER_FILES"] = str(LIB_DIR / "MoltenVK_icd.json")
+if PLATFORM == 'macos':
+    os.environ['VK_DRIVER_FILES'] = str(LIB_DIR / 'MoltenVK_icd.json')
 
 
 # ===============================================================================
 # Util classes
 # ===============================================================================
+
 
 # see https://v4.chriskrycho.com/2015/ctypes-structures-and-dll-exports.html
 class CtypesEnum(IntEnum):
@@ -106,12 +107,11 @@ class CStringArrayType:
     @classmethod
     def from_param(cls, value):
         if not isinstance(value, Iterable) or isinstance(value, (str, bytes)):
-            raise TypeError("Expected a list of strings")
+            raise TypeError('Expected a list of strings')
 
-        encoded = [s.encode("utf-8") for s in value]
+        encoded = [s.encode('utf-8') for s in value]
         bufs = [ctypes.create_string_buffer(s) for s in encoded]
-        arr = (ctypes.c_char_p * len(bufs))(*
-                                            [ctypes.cast(b, ctypes.c_char_p) for b in bufs])
+        arr = (ctypes.c_char_p * len(bufs))(*[ctypes.cast(b, ctypes.c_char_p) for b in bufs])
 
         # keep references alive
         arr._buffers = bufs
@@ -135,25 +135,25 @@ class CStringBuffer:
     """
 
     # -------- allocate an explicit, reusable buffer -----------------
-    def __init__(self, initial: str = "", size: Optional[int] = 64):
+    def __init__(self, initial: str = '', size: Optional[int] = 64):
         if isinstance(initial, Path):
             initial = str(initial)
         if not isinstance(initial, str):
-            raise TypeError("Expected a string")
+            raise TypeError('Expected a string')
 
-        raw = initial.encode("utf-8")
+        raw = initial.encode('utf-8')
         size = (len(raw) + 1) if size is None else size
         if size < len(raw) + 1:
-            raise ValueError("Buffer too small for initial contents")
+            raise ValueError('Buffer too small for initial contents')
 
         self._buf = ctypes.create_string_buffer(size)
-        ctypes.memmove(self._buf, raw, len(raw))   # copy text, keep final NUL
+        ctypes.memmove(self._buf, raw, len(raw))  # copy text, keep final NUL
 
     # -------- read the current value --------------------------------
     @property
     def value(self) -> str:
         """Return the UTF‑8 text currently stored in the buffer."""
-        return self._buf.value.decode("utf‑8")
+        return self._buf.value.decode('utf‑8')
 
     # convenience: expose total capacity (bytes, inc. NUL)
     @property
@@ -175,16 +175,17 @@ class CStringBuffer:
         if isinstance(value, Path):
             value = str(value)
         if isinstance(value, str):
-            buf = ctypes.create_string_buffer(value.encode("utf‑8"))
-            buf._keepalive = buf           # prevent premature GC
+            buf = ctypes.create_string_buffer(value.encode('utf‑8'))
+            buf._keepalive = buf  # prevent premature GC
             return buf
 
-        raise TypeError("Expected CStringBuffer or str")
+        raise TypeError('Expected CStringBuffer or str')
 
 
 # ===============================================================================
 # Out wrapper
 # ===============================================================================
+
 
 class Out:
     _ctype_map = {
@@ -199,7 +200,7 @@ class Out:
         else:
             py_type = type(initial)
             if py_type not in self._ctype_map:
-                raise TypeError(f"Unsupported type: {py_type}")
+                raise TypeError(f'Unsupported type: {py_type}')
             self._ctype = self._ctype_map[py_type]
         self._buffer = self._ctype(initial)
 
@@ -217,7 +218,7 @@ class Out:
     @classmethod
     def from_param(cls, obj):
         if not isinstance(obj, cls):
-            raise TypeError("Expected an Out instance")
+            raise TypeError('Expected an Out instance')
         return ctypes.byref(obj._buffer)
 
     def __format__(self, format_spec):
@@ -231,10 +232,11 @@ class Out:
 # Array wrappers
 # ===============================================================================
 
+
 # HACK: accept None ndarrays as arguments, see https://stackoverflow.com/a/37664693/1595060
 def ndpointer(*args, **kwargs):
-    ndim = kwargs.pop("ndim", None)
-    ncol = kwargs.pop("ncol", None)
+    ndim = kwargs.pop('ndim', None)
+    ncol = kwargs.pop('ncol', None)
     base = ndpointer_(*args, **kwargs)
 
     @classmethod
@@ -242,38 +244,37 @@ def ndpointer(*args, **kwargs):
         if obj is None:
             return obj
         if isinstance(obj, np.ndarray):
-            s = f"array <{obj.dtype}>{obj.shape}"
+            s = f'array <{obj.dtype}>{obj.shape}'
             if ndim and obj.ndim != ndim:
-                raise ValueError(
-                    f"Wrong ndim {obj.ndim} (expected {ndim}) for {s}")
+                raise ValueError(f'Wrong ndim {obj.ndim} (expected {ndim}) for {s}')
             if ncol and ncol > 1 and obj.shape[1] != ncol:
-                raise ValueError(
-                    f"Wrong shape {obj.shape} (expected (*, {ncol})) for {s}")
+                raise ValueError(f'Wrong shape {obj.shape} (expected (*, {ncol})) for {s}')
             out = base.from_param(obj)
         else:
             # NOTE: allow passing ndpointers without change
             out = obj
         return out
 
-    return type(base.__name__, (base,), {"from_param": from_param})
+    return type(base.__name__, (base,), {'from_param': from_param})
 
 
 # ===============================================================================
 # Vec types
 # ===============================================================================
 
+
 class CVectorBase(ctypes.Array):
     _type_ = ctypes.c_float
     _length_ = 0
-    _name_ = ""
+    _name_ = ''
 
     def __new__(cls, *values):
         values = list(values) + [0] * (cls._length_ - len(values))
         return super().__new__(cls, *values[: cls._length_])
 
     def __repr__(self):
-        vals = ", ".join(f"{v:.6g}" for v in self)
-        return f"{self._name_}({vals})"
+        vals = ', '.join(f'{v:.6g}' for v in self)
+        return f'{self._name_}({vals})'
 
 
 VEC_TYPES = []
@@ -284,30 +285,30 @@ def make_vector_type(name, ctype, length):
         name,
         (CVectorBase,),
         {
-            "_type_": ctype,
-            "_length_": length,
-            "_name_": name,
+            '_type_': ctype,
+            '_length_': length,
+            '_name_': name,
         },
     )
     VEC_TYPES.append(t)
     return t
 
 
-vec2 = make_vector_type("vec2", ctypes.c_float, 2)
-vec3 = make_vector_type("vec3", ctypes.c_float, 3)
-vec4 = make_vector_type("vec4", ctypes.c_float, 4)
-dvec2 = make_vector_type("dvec2", ctypes.c_double, 2)
-dvec3 = make_vector_type("dvec3", ctypes.c_double, 3)
-dvec4 = make_vector_type("dvec4", ctypes.c_double, 4)
-cvec4 = make_vector_type("cvec4", ctypes.c_uint8, 4)
-uvec2 = make_vector_type("uvec2", ctypes.c_uint32, 2)
-uvec3 = make_vector_type("uvec3", ctypes.c_uint32, 3)
-uvec4 = make_vector_type("uvec4", ctypes.c_uint32, 4)
-ivec2 = make_vector_type("ivec2", ctypes.c_int32, 2)
-ivec3 = make_vector_type("ivec3", ctypes.c_int32, 3)
-ivec4 = make_vector_type("ivec4", ctypes.c_int32, 4)
-mat3 = make_vector_type("mat3", ctypes.c_int32, 3 * 3)
-mat4 = make_vector_type("mat4", ctypes.c_int32, 4 * 4)
+vec2 = make_vector_type('vec2', ctypes.c_float, 2)
+vec3 = make_vector_type('vec3', ctypes.c_float, 3)
+vec4 = make_vector_type('vec4', ctypes.c_float, 4)
+dvec2 = make_vector_type('dvec2', ctypes.c_double, 2)
+dvec3 = make_vector_type('dvec3', ctypes.c_double, 3)
+dvec4 = make_vector_type('dvec4', ctypes.c_double, 4)
+cvec4 = make_vector_type('cvec4', ctypes.c_uint8, 4)
+uvec2 = make_vector_type('uvec2', ctypes.c_uint32, 2)
+uvec3 = make_vector_type('uvec3', ctypes.c_uint32, 3)
+uvec4 = make_vector_type('uvec4', ctypes.c_uint32, 4)
+ivec2 = make_vector_type('ivec2', ctypes.c_int32, 2)
+ivec3 = make_vector_type('ivec3', ctypes.c_int32, 3)
+ivec4 = make_vector_type('ivec4', ctypes.c_int32, 4)
+mat3 = make_vector_type('mat3', ctypes.c_float, 3 * 3)
+mat4 = make_vector_type('mat4', ctypes.c_float, 4 * 4)
 
 
 # ===============================================================================
